# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Ferramentas",
        "description": "Documentação da API responsável por endpoints relacionados a ferramentas.",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "Servidor principal"
        }
    ],
    "paths": {
...
        "/api/tools": {
            "get": {
                "tags": [
                    "Tools"
                ],
                "summary": "Listar todas as ferramentas",
                "description": "Retorna uma lista de ferramentas com paginação. Opcionalmente pode filtrar por tags.",
                "operationId": "21747b624f411296ff516e7e5e4cd802",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número da página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Número de itens por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "tag",
                        "in": "query",
                        "description": "Filtrar ferramentas por tag",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de ferramentas retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "title": {
                                                        "type": "string"
                                                    },
                                                    "description": {
                                                        "type": "string"
                                                    },
                                                    "link": {
                                                        "type": "string"
                                                    },
                                                    "tags": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "string",
                                                            "example": "php"
                                                        },
                                                        "example": [
                                                            "php",
                                                            "laravel",
                                                            "mysql"
                                                        ]
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tools"
                ],
                "summary": "Criar uma nova ferramenta",
                "description": "Cria uma nova ferramenta com os dados fornecidos.",
                "operationId": "33e623bca5019a757233c3bf6d7924f4",
                "requestBody": {
                    "description": "Dados da nova ferramenta",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "link"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Nova Ferramenta"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descrição da nova ferramenta"
                                    },
                                    "link": {
                                        "type": "string",
                                        "example": "http://exemplo.com"
                                    },
                                    "tags": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "docker",
                                            "jenkins"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tool successfully created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "title": {
                                            "type": "string",
                                            "example": "laravel"
                                        },
                                        "link": {
                                            "type": "string",
                                            "example": "https://www.laravel.com"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "laravel"
                                        },
                                        "tags": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "php"
                                            },
                                            "example": [
                                                "php",
                                                "laravel",
                                                "mysql"
                                            ]
                                        },
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos fornecidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The title is required."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}